refactor: move openapi.d.ts to src/__generated__, rename run -> sync
Committed-By-Agent: claude
