+++ apps/service/src/lib/service-client.ts
@@ -1,5 +1,5 @@
 import createClient from 'openapi-fetch'
-import type { paths } from './openapi.js'
+import type { paths } from '../__generated__/openapi.js'
 
 export function createServiceClient(baseUrl: string) {
   return createClient<paths>({ baseUrl })
